refactor(#120): move hgnc/clinvar/alphagenome constants to per-plugin shared/constants.py
Three plugins' constants relocated in one batched PR per user choice:

  - hgnc (4 constants) -- new shared/ directory created
  - clinvar (4 constants -- 2 URL + 2 label vocabularies)
  - alphagenome (5 default params) -- new shared/ directory created

For clinvar's pathogenicity-label vocabularies (CLINVAR_PATHOGENIC_LABELS,
CLINVAR_BENIGN_LABELS), the three cross-algorithm consumers (algorithms/
ptm, algorithms/psroc, algorithms/annotation) now import directly from
hvantk.skills.clinvar.shared.constants per the user-approved Option A
(make the existing intrinsic coupling explicit; these algorithms
interpret ClinVar's vocabulary by design). algorithms/psroc preserves
its as-aliasing.

Dropped 2 module-import logger.debug dumps (hgnc, clinvar; alphagenome
had none). Updated 11 Python consumers (5 hgnc, 5 clinvar including 3
cross-algorithm, 1 alphagenome) and 3 SKILL.md path references.
Three of the nine PRs closing #120.
