+++ hvantk/skills/hgnc/drift_probe.py
@@ -9,7 +9,7 @@ from datetime import datetime, timezone
 import requests
 
 from hvantk.core.plugin.api import DriftProbeError
-from hvantk.core.constants import HGNC_DOWNLOAD_URL
+from hvantk.skills.hgnc.shared.constants import HGNC_DOWNLOAD_URL
 
 PROBE_VERSION = 1
 _FILENAME = "hgnc_complete_set.txt"
